relocate apiDefine file
